refactor(fuji): split entries.svelte.ts into entries-state and view-state
The single entries.svelte.ts was doing two unrelated things—entry
collection state and view preference state. Split into:
- entries-state.svelte.ts: entriesState + matchesEntrySearch
- view-state.svelte.ts: viewState (view mode, sort, search query)

Updated all 11 consumer imports. Formatting normalized by biome.
