+++ apps/fuji/src/routes/+page.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
 	import EntriesTable from '$lib/components/EntriesTable.svelte';
 	import EntriesTimeline from '$lib/components/EntriesTimeline.svelte';
-	import { entriesState, viewState } from '$lib/entries.svelte';
+	import { entriesState } from '$lib/entries-state.svelte';
+	import { viewState } from '$lib/view-state.svelte';
 </script>
 
 {#if viewState.viewMode === 'table'}
